+++ src/typing_ex/view/page.clj
@@ -8,7 +8,7 @@
    [ring.util.anti-forgery :refer [anti-forgery-field]]
    [typing-ex.plot :refer [scatter]]))
 
-(def ^:private version "5.4.5")
+(def ^:private version "5.4.7-SNAPSHOT")
 
 ;--------------------------------
 (defn- ss
@@ -260,13 +260,17 @@
 
 ;; view of /todays
 (defn todays-act-page [ret login]
+  (def r ret)
   (page
    [:h2 "Typing: Todays"]
    (headline 7)
    [:div {:style "margin-left:1rem;"}
-    [:p "タイピング練習した時刻。10回以上は ... で表示。"]
+    [:p "タイピング練習した時刻。最近練習した人が上で、最近練習した時刻が左。"
+     "10回以上は ... で表示。"]
     [:ol
-     (for [r ret]
+     (for [r (->> ret
+                  (sort-by #(:timestamp (first %)))
+                  reverse)]
        (let [name (-> r first :login)
              tp (-> (mapv #(-> % :timestamp jt/local-time str (subs 0 5)) r))]
          [:li {:style "font-family: monospace;"}
@@ -277,6 +281,16 @@
            name]
           " "
           [:span (abbrev10 tp)]]))]]))
+
+(comment
+  r
+  (first r)
+  (second r)
+  (sort-by #(:timestamp (first %)) r)
+  (->> r
+       (sort-by #(:timestamp (first %)))
+       reverse)
+  :rcf)
 
 (defn sums-page [ret user n]
   (page